Switch from Wand to imageio for GIF rendering
This lets the bot generate smaller GIFs, spend less time on doing so,
and allows specifying the loop count, which is useful for at least
Password and Two Bits.

However, GIFs generated by imageio are currently being mangled by
Discord's media proxy. This is most visible on Simon Says GIFs,
but it can also be seen as an occasional loss of quality on the
antialiased pixels on random frames.

Because of that, deploying this change will have to wait, and that's
why it's on a new branch.