Stop using serenity::client::Context::data
The ShareMap in serenity's Context is a bad way to keep the data -
the locking is not nearly granular enough, and the internal HashMap
reminds me of JavaScript-level engineering. Because of this, the data
will be stored in the event handler struct, which is now passed around
almost as much as the Context.